// src/models/IdData.ts
type IdData = {
    imageUrl: string | null,
    imageName: string | null,
    surname: string | null;
    docNo: string | null;
    givenName: string | null;
    dateOfIssue: string | null;
    dateOfBirth: string | null;
    dateOfExpiry: string | null;
    nationality: string | null;
    sex: string | null;
}

export default IdData;